refactor: Move files into `src` directory
This makes it easier to see application code vs config etc.
