+++ action.yaml
@@ -24,4 +24,4 @@ inputs:
     required: false
 runs:
   using: "node16"
-  main: "index.js"
+  main: "dist/index.js"
