+++ package.json
@@ -3,13 +3,13 @@
   "private": true,
   "version": "0.0.0",
   "description": "A GitHub Action to build and upload Riff-Raff artifacts.",
-  "main": "index.js",
+  "main": "dist/index.js",
   "scripts": {
-    "build": "esbuild index.ts --bundle --platform=node --target=node16 --outfile=index.js",
+    "build": "esbuild src/index.ts --bundle --platform=node --target=node16 --outfile=dist/index.js",
     "test": "jest",
     "tsc": "tsc --noEmit",
-    "lint": "eslint . --ext .ts --no-error-on-unmatched-pattern",
-    "format": "prettier --check \"**/*.ts\""
+    "lint": "eslint src --ext .ts --no-error-on-unmatched-pattern",
+    "format": "prettier --check \"src/**/*.ts\""
   },
   "repository": {
     "type": "git",
