+++ src/deleteRecursively.ts
